Extract Jaeger::Exporter::SpanEncoder class.
